Make crate more compatible with firewheel / bevy_seedling (#20)
* Impl Send for ReflectionEffectParams

* Use preallocated pointer buffer

* More random stuff I need

* Some more

* Add partial eq

* Add ReflectionEffectIR(

* Sync

* Add derives

* Move constructor

* Add firewheel feature

* Fix gate

* Remove unnecessary indirection

* Revert borrowing changes in demo and test

* Fix compiler errors

* Move stuff to u32

* Fix tests

* Add std to firewheel

* Fix lints

* Apply suggestions from code review


* Use From

* Fix lints

* Improve API soundness

* Make things mut

* Fix mut

* Add PartialEq

* Remove firewheel feature

* Add default

* Remove mut

* Remove ref

* Remove ref

* Remove ref

* Add identity matrix

* Remove ref

* Fix oops

* Remove ref

---------

+++ audionimbus/src/audio_settings.rs
@@ -1,12 +1,12 @@
 /// Global settings for audio signal processing.
-#[derive(Debug)]
+#[derive(Debug, Copy, Clone)]
 pub struct AudioSettings {
     /// Sampling rate, in Hz.
-    pub sampling_rate: usize,
+    pub sampling_rate: u32,
 
     /// Frame size, in samples.
     /// Independent of number of channels.
-    pub frame_size: usize,
+    pub frame_size: u32,
 }
 
 impl Default for AudioSettings {
